REF: separate out dispatch-centric ops functions (#28624)
